Implement review feedback
- Pull tests into seperate modules to fix deploy issue
- Add test for host on two networks
- Fix typo in service name
